requirements.txt
PyMuPDF>=1.23.0
matplotlib>=3.8.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
markdown>=3.4.0
requests>=2.31.0
